Improving relay creation (not finished yet and not fully support by dedicated anyway)
git-svn-id: http://maniaplanet-dedicated-manager.googlecode.com/svn/trunk@20 7f24f777-0cea-ae19-a2bf-98aad5406375

+++ libraries/DedicatedManager/Services/ConfigFileService.php
@@ -83,7 +83,7 @@ class ConfigFileService extends DedicatedFileService
 		$config->ladderServerLimitMax = (int) $configObj->server_options->ladder_serverlimit_max;
 		$config->ladderServerLimitMin = (int) $configObj->server_options->ladder_serverlimit_min;
 		$config->nextCallVoteTimeOut = (int) $configObj->server_options->callvote_timeout;
-		$config->callVoteRatio = (double) $configObj->server_options->callvote_ratio;
+		$config->callVoteRatio = (float) $configObj->server_options->callvote_ratio;
 		$config->allowMapDownload = self::toBool($configObj->server_options->allow_map_download);
 		$config->autoSaveReplays = self::toBool($configObj->server_options->autosave_replays);
 		$config->autoSaveValidationReplays = self::toBool($configObj->server_options->autosave_validation_replays);
